Factor out ARROW_ROOT_DEFAULT, fix tests
