+++ dev/archery/archery/integration/tester_java.py
@@ -20,7 +20,8 @@ import os
 import subprocess
 
 from .tester import Tester
-from .util import run_cmd, ARROW_ROOT_DEFAULT, log
+from .util import run_cmd, log
+from ..utils.source import ARROW_ROOT_DEFAULT
 
 
 def load_version_from_pom():
